commit
commit

+++ SpringBoot-Rest/src/main/java/com/Abhijith/SpringBoot/Rest/service/JobService.java
@@ -1,6 +1,7 @@
 package com.Abhijith.SpringBoot.Rest.service;
 
 
+import com.Abhijith.SpringBoot.Rest.JobNotFoundException;
 import com.Abhijith.SpringBoot.Rest.Repo.JobRepo;
 import com.Abhijith.SpringBoot.Rest.model.JobPost;
 import org.springframework.stereotype.Service;
@@ -26,13 +27,27 @@ public class JobService {
     public JobPost getJob(int id) {
         return repo.getJob(id);
     }
-   public JobPost updateJob(JobPost jobPost) {
-       return repo.updateJob(jobPost);
+   public JobPost updateJob(JobPost jobPost,int id) {
+       return repo.updateJob(jobPost,id);
    }
 
-    public void deleteJob(int id) {
+    public JobPost deleteJob(int id) {
+        // Fetch job details by id
+        JobPost jobPost = repo.findById(id);
+
+        // If job not found, throw a custom exception
+        if (jobPost == null) {
+             throw new JobNotFoundException("Service issue: Something went wrong. Job with ID " + id + " not found.");
+
+        }
+
+        // Perform the actual deletion from the repository
         repo.deleteJob(id);
+
+        // Return the deleted job
+        return jobPost;
     }
+
 
 
 }
